pyproject.toml
[project]
name = "streamrag"
version = "0.1.0"
description = "Streaming RAG for anomaly detection"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "numpy>=2.3.3",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
    "ollama>=0.3.0",
]
